+++ crates/ferritin-plms/src/ligandmpnn/mod.rs
@@ -3,7 +3,7 @@
 //! - utilities to convert sequence formats (PDB; mmcif) to ML-ready tensors.
 //! - CLI to handle the above.
 //!
-mod commands;
+// mod commands;
 pub mod configs;
 pub mod model;
 pub mod proteinfeatures;
